Remove sticky header on mobile in landscape mode
Well, that wasn't easy as there is no reliable way to make sure that
a mobile device is in landscape mode. I had to come up with a media
query that would not trigger on desktops in undesirable scenarios. I
hope that the navigator.userAgentData.mobile property one day gets the
necessary cross-browser support because it would make tasks such as
this trivial.